lesson_8/docker/requirements.functions.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
boto3==1.17.83
botocore==1.20.83
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.7
frozenlist==1.2.0
idna==2.10
jmespath==0.10.0
mongo==0.2.0
multidict==5.2.0
pika==1.2.0
psycopg2-binary==2.9.2
pymongo==3.11.4
PyMySQL==1.0.2
python-dateutil==2.8.1
redis==3.5.3
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
tika==1.24
typing_extensions==4.0.0
urllib3==1.26.5
yarl==1.7.2
opentelemetry-api==0.15b0
opentelemetry-exporter-jaeger==0.15b0
opentelemetry-sdk==0.15b0
opentracing==2.4.0
jaeger-client==4.3.0